dev (next js): move utils to library
